Change directory structure on rhcsa8 release branch
Starting with the release branch for rhcsa8, the RHCSA and RHCE certs will have
their own directories without version numbers.  Instead, as new versions of the
exams are released, we can create new release branches for them.  The release
branch GitHub URLs are what should be linked in external docs for specific
versions.